requirements.txt
# =============================================
# Fish Disease Detection - Dependencies
# =============================================
# Python 3.11 | Flask 2.x | TensorFlow/Keras
# =============================================

# --- Web Framework ---
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3

# --- Database ---
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# --- Deep Learning ---
tensorflow==2.13.1
keras==2.13.1

# --- Image Processing ---
Pillow==10.1.0
opencv-python==4.8.1.78

# --- Data Science ---
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
scikit-learn==1.3.2

# --- Utilities ---
python-dotenv==1.0.0
tqdm==4.66.1

# --- Production Server ---
gunicorn==21.2.0
